prepit: Use os.path.
os.path.join is much safer than assuming that URLs from environment
variables do not contain a trailing slash.  Since we are always in
control of the suffix, there is no reason to not use it here.

The same is true for using os.path to parse file names.  I ran into a
case where I was using a different file extension for postgres files.
Using os.path fixes that specific problem as well.